+++ outputs/scripting.sh
@@ -1,14 +1,14 @@
 #!/bin/bash
 
 # Source definitions
-project=$(dirname "$0")
-. $project/defs.sh
+project="$(dirname "$0")"
+. "$project/defs.sh"
 
-rm -rf $raw_outputs/scripting
-. $project/init_spack.sh
+rm -rf "${raw_outputs:?}/scripting"
+. "$project/init_spack.sh"
 
 spack install gcc@8.4.0
-spack compiler find `spack location -i gcc@8.4.0`
+spack compiler find "$(spack location -i gcc@8.4.0)"
 
 example scripting/setup           "spack uninstall -ay"
 example scripting/setup           "spack compiler rm gcc@8.4.0"
@@ -24,7 +24,7 @@ echo "exit()
 
 fake_example scripting/edit '$EDITOR find_exclude.py' "/bin/true"
 
-cat <<EOF | tee ${PROJECT}/raw/0.find_exclude.py.example find_exclude.py
+cat <<EOF | tee "$PROJECT/raw/0.find_exclude.py.example" find_exclude.py
 from spack.spec import Spec
 import spack.store
 import spack.cmd
@@ -41,16 +41,16 @@ EOF
 
 example scripting/find-exclude-1 "spack python find_exclude.py %gcc ^mpich"
 
-EXAMPLE1="${PROJECT}/raw/1.find_exclude.py.example"
-echo "#!/usr/bin/env spack python" > $EXAMPLE1
-cat find_exclude.py >> $EXAMPLE1
-cp $EXAMPLE1 find_exclude.py
+EXAMPLE1="$PROJECT/raw/1.find_exclude.py.example"
+echo "#!/usr/bin/env spack-python" > "$EXAMPLE1"
+cat find_exclude.py >> "$EXAMPLE1"
+cp "$EXAMPLE1" find_exclude.py
 
 example scripting/find-exclude-2 "chmod u+x find_exclude.py"
 example scripting/find-exclude-2 "./find_exclude.py %gcc ^mpich"
 
-EXAMPLE2="${PROJECT}/raw/2.find_exclude.py.example"
-sed s'|spack python|spack-python|' find_exclude.py > $EXAMPLE2
-cp $EXAMPLE2 find_exclude.py
+EXAMPLE2="$PROJECT/raw/2.find_exclude.py.example"
+sed s'|spack python|spack-python|' find_exclude.py > "$EXAMPLE2"
+cp "$EXAMPLE2" find_exclude.py
 
 example scripting/find-exclude-3 "./find_exclude.py %gcc ^mpich"
